Bug fix
There is an issue when using the root directory as an argument, like:
"sizeof /", the total size displayed will be around like hundreds of
terabytes greater than the size of the storage which is crazy, I only
have 512G in my nvme storage and it displays like 100+ terabytes when I
check the size of the root directory, this also happens in the "du"
command when you do "du --apparent-size -h", the same bug happens with
the du command lol.

I figured out that this was caused by the file in /prov/kcore file, I'm
not sure what it does but I think it has something to do with the kernel
because when you inspect the file using the command "file" with root
privileges, it says stuff like 'BOOT_IMAGE=/boot/vmlinuz-linux ...', but
I have no fucking idea why it is so big, all I heard is that it is a
pseudo-file and it is not really a file or smth, same goes with all the
other 'files' in the /proc directory, it kinda makes sense because these
'files' don't seem to take up blocks of space.

So I decided that the fix is to just ignore files that has a block size
of <= 0, I also separated the conditions of (buf.st_size <= 0 &&
access(file_name, F_OK) != 0) because of the files in the /proc
directory.

I also switched the type of "size" to be size_t instead of double, not
sure if this makes things faster or leaner or both but I think it is
more appropriate to use size_t as the type for bytes.

+++ include/lib.h
@@ -1,12 +1,13 @@
 #ifndef LIB_H
 #define LIB_H
+#include <stdlib.h>
 #include <sys/stat.h>
 
 // slices a file name if the file name contains directories of its path,
 // e.g., "src/main.c", the function should return "main.c"
 char *truncate_file_name(char filename[]);
 char *concat(const char *s1, const char *s2);
-void process_output(char *file_name, double size, int truncate_flag,
+void process_output(char *file_name, size_t size, int truncate_flag,
 					int is_dir);
 int evaluate_st_mode(struct stat buf, char *file_name);
 
@@ -14,9 +15,9 @@ int evaluate_st_mode(struct stat buf, char *file_name);
 // is lesser than a KiB, it should only display size in bytes, or if
 // the file size is greater than a GiB, it should display all sizes in
 // bytes, KiB, MiB, GiB
-void display_file_output_full(char *final_file_name, double size);
-void display_file_output_lesser(char *final_file_name, double size);
-void display_dir_output(char *final_file_name, double size);
+void display_file_output_full(char *final_file_name, size_t size);
+void display_file_output_lesser(char *final_file_name, size_t size);
+void display_dir_output(char *final_file_name, size_t size);
 
 void process_directory(char *dir_name);
 void process_file(char *file_name, int truncate_flag, int is_dir);
